WIP: Change login page to new UI with Bulma
