+++ src/components/Header/Header.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import './Header.scss';
+
+const Header = () => (
+  <nav className="navbar is-info">
+    <div className="navbar-brand">
+      <span className="navbar-item navbar-item-single">
+        Simple Messenger
+      </span>
+    </div>
+  </nav>
+);
+
+export default Header;
