+++ src/components/Header/Header.scss
@@ -1,0 +1,3 @@
+.navbar-brand {
+    justify-content: center;
+}
