+++ src/components/Hero/Hero.js
@@ -1,0 +1,20 @@
+import React from 'react';
+import './Hero.scss';
+
+const Hero = () => (
+  <section class="hero">
+    <div class="hero-body">
+      <div class="container">
+        <h1 class="title">
+          Sign In to talk with 
+          admin@admin.com
+        </h1>
+        <h2 class="subtitle">
+          Para fazer login, preencha seu e-mail abaixo e entre no link que enviaremos para você.
+        </h2>
+      </div>
+    </div>
+  </section>
+);
+
+export default Hero;
