+++ src/containers/Login/Login.css
@@ -1,8 +1,8 @@
 .container{
-    display: flex;
+    /* display: flex;
     align-items: center;
     justify-content: center;
-    width: 100%;
+    width: 100%; */
 }
 
 .container-home {
