+++ src/index.css
@@ -7,13 +7,13 @@ body {
 	font-size: 1.5em;
   color: #333;
   margin: 0px;
-  background-color: dodgerblue;
+  background-color: white;
 }
 
 .container {
-	max-width: 100%;
-	margin: 0 auto;
-	display: flex;
+	/* max-width: 100%; */
+	/* margin: 0 auto; */
+	/* display: flex; */
 	/* border: 1px solid #ccc; */
 }
